:truck: (history) rework folder organization
